services/worker/requirements.txt
celery==5.3.6
redis==5.0.1
cryptography==41.0.7
pyOpenSSL==23.3.0
python-dateutil==2.8.2
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
prometheus-client==0.19.0
